最終

+++ students/FIREBAR/WeaponManager.h
@@ -20,7 +20,7 @@ struct Weapons
 	std::string name;
 
 	//武器のダメージ
-	float damage;
+	int damage;
 
 	//武器の射程距離
 	float range;
@@ -85,7 +85,7 @@ public:
 	/// <param name="num">武器の番号
 	/// ０：弓、１：刀、２：斧、３：魔法</param>
 	/// <returns>対応する武器のステータス</returns>
-	float GetWeaponDamage(int num) { return weapons[num].damage; }
+	int GetWeaponDamage(int num) { return weapons[num].damage; }
 
 	/// <summary>
 	///	攻撃力増加
